Dave Culp:
I've just completed a Lockheed F-80C Shooting Star.

    http://home.comcast.net/~davidculp2/F80C.tar.gz

screenshot:

    http://home.comcast.net/~davidculp2/P80C-001.jpg

The exterior model was originally done by Massimo Taccoli, who has done
several very nice models for MSFS, and has given me permission to convert it
for use in FlightGear and release the resulting model under GPL.  The FDM is
JSBSim, based on Aero-Matic output with a couple tweaks.  It runs here in CVS
FlightGear, and should run in 0.9.8 and later?

It's a no-brainer to fly.  Engine is running at startup, and flaps are set at
half (takeoff position).  Check out the gunsight (you have to turn it on by
clicking on the on/off switch).  It has auto-depression for g-load.