Anders Gidenstam:
I recently noticed that the F80C has a tendency to flip upside down on the
runway and then keep rolling along upside down hanging from its wheels.

This diff adds contact points at the wingtip fuel tanks and at the nose
and tail. I also moved the visual reference point 185" back to match the
origo of the 3d-model file. I think that was better than changing the
offset of the 3d-model since the aircraft following views in FlightGear
centers on the VRP.